refactor: update calificacion data type in Restaurante class

+++ src/com/krakedev/Restaurante.java
@@ -3,5 +3,5 @@ package com.krakedev;
 public class Restaurante {
     public String nombre;
     public String direccion;
-    public float calificacion;
+    public double calificacion;
 }
